requirements.txt
# ── core ──────────────────────────────────────────────
python-dotenv==1.0.1
googlemaps==4.10.0
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
snowflake-connector-python==3.10.1
snowflake-snowpark-python==1.20.0\
scikit-learn==1.5.1
umap-learn==0.5.6
hdbscan==0.8.38.post1

bertopic==0.16.4
sentence-transformers==3.0.1
spacy==3.7.5

streamlit==1.36.0
folium==0.17.0
plotly==5.22.0